Issue #1757 tests merged into WebsocketTest.java
